requirements.txt
# --- Core FastAPI ---
fastapi==0.128.0
uvicorn==0.39.0
starlette==0.49.3
python-multipart==0.0.20
pydantic==2.12.5
pydantic_core==2.41.5
typing_extensions==4.15.0
anyio==4.12.0
h11==0.16.0

# --- Auth / Security ---
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
PyJWT==2.10.1
pycryptodome==3.23.0
python-dotenv==1.0.0

# --- HTTP / Networking ---
requests==2.32.5
httpx==0.28.1
dnspython==2.6.1
beautifulsoup4==4.12.3
aiohttp==3.10.11
aiodns==3.6.1
aiohappyeyeballs==2.6.1
multidict==6.7.0
yarl==1.22.0
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.2
certifi==2026.1.4

# --- Database ---
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
greenlet==3.2.4

# --- Storage / MinIO ---
minio==7.2.20

# --- Reporting / PDF ---
fpdf2==2.7.9
Pillow==11.3.0
fonttools==4.60.2

# --- Data handling (used by scanners/reports) ---
numpy==2.0.2
pandas==2.3.3
openpyxl==3.1.5

# --- Observability ---
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-semantic-conventions==0.59b0

# --- Background Jobs (Celery + Redis) ---
celery==5.4.0
redis==5.2.1

# --- Utils ---
rich==14.2.0
tenacity==9.1.2
PyYAML==6.0.3
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
six==1.17.0
zipp==3.23.0
importlib_metadata==8.7.1
